feat(ws)!: retire terminal.codex.candidate.persisted writer - accept-and-ignore with debug log
One writer per codex identity fact (campaign §2.3.2): the server-side
rollout locator is authoritative. The protocol variant stays for
wire-compat with the frozen client; the write path, its four guards,
and both candidate integration tests are deleted (superseded by
codex_candidate_inert.rs and codex_locator_activity.rs).
